requirements.txt
# ── Web / API stack ───────────────────────────────────────────────
fastapi>=0.110,<0.140
uvicorn[standard]>=0.27,<0.50
python-multipart>=0.0.9
pydantic>=2.5

# ── Desktop window ────────────────────────────────────────────────
pywebview>=4.0

# ── Vision / ML (non-torch) ───────────────────────────────────────
opencv-python-headless>=4.8
onnxruntime>=1.17
scikit-learn>=1.3
scipy>=1.10
numpy>=1.24,<3.0
Pillow>=10.0

# ── CLIP / torch installed separately by Start scripts ───────────
# (torch index differs between Windows CPU and Mac/Apple-Silicon)
# torch>=2.1
# torchvision
# clip @ git+https://github.com/openai/CLIP.git

# ── Utilities ─────────────────────────────────────────────────────
tqdm>=4.65
joblib>=1.3
ftfy
regex
piexif>=1.1
fpdf2>=2.7
requests>=2.28
